+++ app/src/main/AndroidManifest.xml
@@ -12,10 +12,10 @@
         android:theme="@style/Theme.StudioWedding"
         tools:targetApi="31">
         <activity
-            android:name=".UpdateEmployeeActivity"
+            android:name=".view.activity.employee.UpdateEmployeeActivity"
             android:exported="false" />
         <activity
-            android:name=".AddEmployeeActivity"
+            android:name=".view.activity.employee.AddEmployeeActivity"
             android:exported="false" />
         <activity
             android:name=".SplashActivity"
