--- app/src/main/java/com/example/studiowedding/view/activity/Employee/AddEmployeeActivity.java
@@ -1,14 +1,0 @@
-package com.example.studiowedding.view.activity.Employee;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class AddEmployeeActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_employee);
-    }
-}
