+++ app/src/main/java/com/example/studiowedding/view/activity/employee/AddEmployeeActivity.java
@@ -1,4 +1,4 @@
-package com.example.studiowedding.view.activity.Employee;
+package com.example.studiowedding.view.activity.employee;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import com.example.studiowedding.R;
 
-public class UpdateEmployeeActivity extends AppCompatActivity {
+public class AddEmployeeActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_update_employee);
+        setContentView(R.layout.activity_add_employee);
     }
 }
