+++ app/src/main/java/com/example/studiowedding/view/activity/employee/UpdateEmployeeActivity.java
@@ -1,0 +1,16 @@
+package com.example.studiowedding.view.activity.employee;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+import com.example.studiowedding.R;
+
+public class UpdateEmployeeActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_update_employee);
+    }
+}
