+++ app/src/main/res/drawable/btn_add_employee.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="69dp"
+    android:height="69dp"
+    android:viewportWidth="69"
+    android:viewportHeight="69">
+  <path
+      android:pathData="M34.5,34.5m-34.5,0a34.5,34.5 0,1 1,69 0a34.5,34.5 0,1 1,-69 0"
+      android:fillColor="#D9D9D9"/>
+  <path
+      android:pathData="M34.75,26C35.13,26 35.443,26.282 35.493,26.648L35.5,26.75L35.501,34H42.754C43.168,34 43.504,34.336 43.504,34.75C43.504,35.13 43.222,35.444 42.856,35.493L42.754,35.5H35.501L35.503,42.749C35.503,43.163 35.168,43.499 34.753,43.499C34.374,43.499 34.06,43.217 34.01,42.851L34.003,42.749L34.001,35.5H26.752C26.338,35.5 26.002,35.164 26.002,34.75C26.002,34.37 26.284,34.056 26.65,34.007L26.752,34H34.001L34,26.75C34,26.336 34.336,26 34.75,26Z"
+      android:fillColor="#212121"/>
+</vector>
